Refactor & fix Mojang API Error Add max task limit to fix Mojang API Error Fixes #5
